Store metrics separately to avoid double-borrow
